+++ src/common/types.ts
@@ -1,4 +1,12 @@
-export type TransportTest = { filePath: string; title?: string; tags?: Mocha.GrepTag[] };
+export interface GrepTagObject {
+  tag: string;
+  info?: string[];
+}
+
+export type GrepTagSimple = string;
+export type GrepTag = GrepTagObject | GrepTagSimple;
+
+export type TransportTest = { filePath: string; title?: string; tags?: GrepTag[] };
 
 export type ParsedSpecs = {
   parentFolder?: string;
