+++ src/mocha/index.ts
@@ -1,11 +1,7 @@
 // eslint-disable-next-line @typescript-eslint/no-unused-vars,@typescript-eslint/no-namespace
 namespace Mocha {
-  export interface GrepTagObject {
-    tag: string;
-    info?: string[];
-  }
-  export type GrepTagSimple = string;
-  export type GrepTag = GrepTagObject | GrepTagSimple;
+  export type GrepTagObject = import('../common/types').GrepTagObject;
+  export type GrepTag = import('../common/types').GrepTag;
 
   export interface Test {
     tags?: GrepTagObject[];
